requirements.txt
aioredis==2.0.0b1
asgiref==3.4.1
async-timeout==3.0.1
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
fastapi==0.66.0
fastapi-cache2==0.1.4
h11==0.12.0
httptools==0.2.0
idna==3.2
pydantic==1.8.2
python-dateutil==2.8.2
python-dotenv==0.18.0
PyYAML==5.4.1
requests==2.26.0
six==1.16.0
starlette==0.14.2
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.14.0
watchgod==0.7
websockets==9.1
